Moved the factory implementation into a separate module.
